requirements.txt
pySide6==6.5.1
h5py
pillow
numpy